wrangled asoundrc; changed XDG_CONFIG_HOME to ~/.local/etc; ... so on and so forth.

+++ xinitrc.d/compton.sh
@@ -1,13 +1,15 @@
 #!/usr/bin/env bash
 
-#
+##
 ## Start compton running as a daemon
-#
+##
 
-[[ -f ${XDG_RUNTIME_DIR}/compton@${DISPLAY}.pid ]] && {
+if test -f "${XDG_RUNTIME_DIR}/compton@${DISPLAY}.pid" \
+&& test -r "${XDG_RUNTIME_DIR}/compton@${DISPLAY}.pid"
+then
   kill -15 "$(< "${XDG_RUNTIME_DIR}/compton@${DISPLAY}.pid")"
   rm "${XDG_RUNTIME_DIR}/compton@${DISPLAY}.pid"
-}
+fi
 
 compton     \
   --daemon  \
